# BE/api-spec.yml
openapi: "3.0.2"
info:
  title: Anti Retake API
  version: '1.2'
servers:
  - url: https://antiretake.xyz/api
paths:
...
  /users/{userId}/folders:
    get:
      summary: Get folder list of a user by UserID.
      tags:
        - Profile
      operationId: API-013
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Folder list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderList"
...
  /classes/{classId}/folders:
    get:
      summary: Get folders in class by class ID.
      tags:
        - Class
      operationId: API-035
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Folders of class.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderList"
  /classes/{classId}/members:
    get:
      summary: Get members in class by class ID.
      tags:
        - Class
      operationId: API-036
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members of class.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberList"
...
  /classes/{classId}:
    delete:
      summary: Delete class by class Id.
      tags:
        - Class
      operationId: API-040
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Class deleted successful.

    put:
      summary: Edit a class.
      tags:
        - Class
      operationId: API-039
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassEdit"
      responses:
        "201":
          description: Class updated successful.
...
  /folders/{folderId}/add-study-set-to-folder:
    post:
      summary: Add study set to folder.
      tags:
        - Folder
      operationId: API-076
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: folderId
          required: true
          schema:
            type: string
        - in: query
          name: studySetId
          schema:
            type: string
      responses:
        "200":
          description: Add study set to folder successful.
...
  /forum/questions/{questionId}/:
    get:
      summary: Get question by question Id.
      tags:
        - Forum
      operationId: API-056
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: question.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

    put:
      summary: Edit a question.
      tags:
        - Forum
      operationId: API-064
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionEdit"
      responses:
        "201":
          description: Edit a question.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

    delete:
      summary: Delete question by quesiton Id.
      tags:
        - Forum
      operationId: API-062
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Question deleted successful.
...
  /forum/comments/{commentId}:
    put:
      summary: Edit a comment.
      tags:
        - Forum
      operationId: API-060
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentEdit"
      responses:
        "201":
          description: Edit a comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

    delete:
      summary: Delete comment by comment Id.
      tags:
        - Forum
      operationId: API-061
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Comment deleted successful.
...
  /forum/questions/{questionId}/comments/{commentId}/vote:
    delete:
      summary: Delete vote comment.
      tags:
        - Forum
      operationId: API-081
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: string
        - in: path
          name: commentId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Delete successfully.
components:
  schemas:
...
    FolderCardInfo:
      type: object
      properties:
        id:
          type: string
        numOfSets:
          type: number
        name:
          type: string
        createdTime:
          type: string
    FolderList:
      type: array
      items:
        $ref: "#/components/schemas/FolderCardInfo"
...
    MemberInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avaUrl:
          type: string
        accessDay:
          type: string
        role:
          type: string
    MemberList:
      type: array
      items:
        $ref: "#/components/schemas/MemberInfo"
...
    ClassEdit:
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        description:
          type: string
        accessModifier:
          type: string
...
    Comment:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        ownerName:
          type: string
        ownerAva:
          type: string
        createdTime:
          type: string
        numOfUpvotes:
          type: number
        numOfDownvotes:
          type: number
        listOfVotes:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
        content:
          type: string
        listOfComments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        replyToQuestion:
          type: string
        replyToComment:
          type: string
    Vote:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        votedTime:
          type: string
        type:
          type: string
    Question:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdTime:
          type: string
        ownerId:
          type: string
        ownerAvaUrl:
          type: string
        ownerName:
          type: string
        content:
          type: string
        numOfUpvotes:
          type: number
        numOfDownvotes:
          type: number
        listOfVotes:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
        numOfComments:
          type: number
        textbooks:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        schools:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    CommentEdit:
      type: object
      properties:
        content:
          type: string
    QuestionEdit:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        textbooks:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        schools:
          type: array
          items:
            type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ar_Auth